requirements.txt
Flask==1.1.1
flask-sqlalchemy==2.4.1
python-dotenv==0.10.3
flask-marshmallow==0.10.1
Flask-Bcrypt==0.7.1
Flask-Script==2.0.6
Flask-Migrate==2.5.2
psycopg2==2.8.4
flask-restful==0.3.7
flask-httpauth==3.3.0
flask-restful-swagger==0.20.1
requests==2.22.0